minimized popup for glucose / still needs fixing

+++ core/src/main/java/cellcorp/gameofcells/objects/Cell.java
@@ -541,12 +541,8 @@ public class Cell {
         return MAX_ATP;
     }
 
-    /**
-     * Get Cell Height
-     * @return
-     */
-    public int getHeight() {
-        return (int) cellSize;
+    public float getHeight() {
+        return cellSize;
     }
 
     /**
